Moving all plotting functions to a new parcels/plotting.py file
